Fixed Service Loading problem in the Arquillian Extension
